bubble-sort/requirements.txt
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
matplotlib==3.8.2  # Visualization and plotting
pytest==7.4.3  # Testing framework
